# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/tables/{tableId}/columns": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Column List",
        "operationId": "column-list",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ColumnList"
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        },
        "description": "Read project details",
        "tags": [
          "Meta"
        ]
      },
      "post": {
        "summary": "Column create",
        "operationId": "column-create",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Column"
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
    "/tables/{tableId}/columns/{columnId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "columnId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Column Read",
        "operationId": "column-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        },
        "description": "Read project details",
        "tags": [
          "Meta"
        ]
      },
      "put": {
        "summary": "Column Update",
        "operationId": "column-update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Column"
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "delete": {
        "summary": "",
        "operationId": "column-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
...
    "/views/{viewId}/columns": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "viewId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "view-column-list",
        "responses": {
          "200": {}
        },
        "tags": [
          "Meta"
        ]
      },
      "post": {
        "summary": "",
        "operationId": "view-column-create",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
    "/views/{viewId}/columns/{columnId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "viewId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "columnId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "view-column-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "put": {
        "summary": "",
        "operationId": "view-column-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
    "/audits/comments": {
      "parameters": [],
      "get": {
        "summary": "",
        "operationId": "comment-list",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "": {
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "row_id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fk_model_id",
            "required": true
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "comments_only"
          }
        ],
        "tags": [
          "Meta"
        ]
      },
      "post": {
        "summary": "",
        "operationId": "comment-row",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "row_id": {
                    "type": "string"
                  },
                  "fk_model_id": {
                    "type": "string"
                  },
                  "comment": {
                    "type": "string"
                  }
                },
                "required": [
                  "row_id",
                  "fk_model_id",
                  "comment"
                ]
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
    "/audits/comments/count": {
      "parameters": [],
      "get": {
        "summary": "",
        "operationId": "comment-count",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "": {
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "",
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "ids",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fk_model_id",
            "required": true
          }
        ],
        "tags": [
          "Meta"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Column": {
        "title": "Column",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "base_id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "uidt": {
            "type": "string"
          },
          "dt": {
            "type": "string"
          },
          "np": {
            "type": "string"
          },
          "ns": {
            "type": "string"
          },
          "clen": {
            "type": [
              "string",
              "integer"
            ]
          },
          "cop": {
            "type": "string"
          },
          "pk": {
            "type": "boolean"
          },
          "pv": {
            "type": "boolean"
          },
          "rqd": {
            "type": "boolean"
          },
          "un": {
            "type": "boolean"
          },
          "ct": {
            "type": "string"
          },
          "ai": {
            "type": "boolean"
          },
          "unique": {
            "type": "boolean"
          },
          "cdf": {
            "type": "string"
          },
          "cc": {
            "type": "string"
          },
          "csn": {
            "type": "string"
          },
          "dtx": {
            "type": "string"
          },
          "dtxp": {
            "type": "string"
          },
          "dtxs": {
            "type": "string"
          },
          "au": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "colOptions": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LinkToAnotherRecord"
              },
              {
                "$ref": "#/components/schemas/Formula"
              },
              {
                "$ref": "#/components/schemas/Rollup"
              },
              {
                "$ref": "#/components/schemas/Lookup"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SelectOptions"
                }
              }
            ]
          },
          "cn": {
            "type": "string"
          },
          "_cn": {
            "type": "string"
          }
        }
      },
      "ColumnList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "columns": {
            "type": "object",
            "required": [
              "list"
            ],
            "properties": {
              "list": {
                "type": "array",
                "uniqueItems": true,
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        },
        "required": [
          "columns"
        ]
      },
      "LinkToAnotherRecord": {
        "title": "LinkToAnotherRecord",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          },
          "fk_column_id": {
            "type": "string"
          },
          "fk_child_column_id": {
            "type": "string"
          },
          "fk_parent_column_id": {
            "type": "string"
          },
          "fk_mm_model_id": {
            "type": "string"
          },
          "fk_related_model_id": {
            "type": "string"
          },
          "fk_mm_child_column_id": {
            "type": "string"
          },
          "fk_mm_parent_column_id": {
            "type": "string"
          },
          "ur": {
            "type": "string"
          },
          "dr": {
            "type": "string"
          },
          "fk_index_name": {
            "type": "string"
          },
          "deleted": {
            "type": "string"
          },
          "order": {
            "type": "string"
          }
        }
      },
      "Lookup": {
        "title": "Lookup",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          },
          "fk_column_id": {
            "type": "string"
          },
          "fk_relation_column_id": {
            "type": "string"
          },
          "fk_lookup_column_id": {
            "type": "string"
          },
          "deleted": {
            "type": "string"
          },
          "order": {
            "type": "string"
          }
        }
      },
      "Rollup": {
        "title": "Rollup",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          },
          "fk_column_id": {
            "type": "string"
          },
          "fk_relation_column_id": {
            "type": "string"
          },
          "fk_rollup_column_id": {
            "type": "string"
          },
          "rollup_function": {
            "type": "string"
          },
          "deleted": {
            "type": "string"
          },
          "order": {
            "type": "string"
          }
        }
      },
      "Formula": {
        "title": "Lookup",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          },
          "fk_column_id": {
            "type": "string"
          },
          "formula": {
            "type": "string"
          },
          "deleted": {
            "type": "string"
          },
          "order": {
            "type": "string"
          }
        }
      },
      "SelectOptions": {
        "title": "SelectOptions",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          },
          "fk_column_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
...
    },
...
    "responses": {
...
      "ColumnList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ColumnList"
            }
          }
        }
      },
...
    }
  }
}